Use a hash map to store the `ArgparseOption` instance so to reduce the time complexity of finding an individual option
To reduce the time complexity of finding a argument using its short or long
name (depending on what's given) we generate a hash code using the short or
the long name (long name is preferred) then we map the option with the
generated hash code. By doing this in future whenever we will need to check
the other attributes of the `ArgparseOption` instance and the option name is
given to us we can easily find it in constant time.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>