Parse arguments given over the command-line
Parse the arguments given over the command-line and set their values
irrespective of their shorter or longer version name is used.

Use a longToShortOptionsMap or shortToLongOptionsMap to quickly find the
longer or the shorter version of the command used over the command-line.

Note, in our parse function while parsing the argument name we assume the
prefix to be of length 2 which is not true for the current version of the
ArgparseOption class. So keep a note for modifying the class ArgparseOption
in future to only have a default prefix which is "--" not any other prefix
given by the user.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>