Check if multiple arguments from a group are present over the command-line
The main purpose of a group type argument is to forbid the presence of
another argument from the same set a argument belongs to. Thus, we have to
reflect the same behaviour in our program by adding conditional checks to
forbid the presence of another argument from a group. We does so by
immediately halting the program as soon we find that multiple arguments from
a group are present over the command-line.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>