Add wrapText function for text wrapping
This commit adds a new function called wrapText which takes a string of text
and wraps it to a specified width.

It takes several optional parameters to handle whitespaces, newline
characters, special characters and indentation.

This function will be useful for formatting text in a console or any other
application that requires text wrapping.

It also includes handling special characters for console only

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>