Add build & test maven workflow to build and test the `Argparse` package
This commit introduces a maven workflow for GitHub actions to build and test
the package.

Also, this commit refactors the main package files into one single "Argparse"
module to make it a single file package to be used in multiple projects.

This commit also introduce "google-java-format" styles to make the project
conform to the google style guide for java.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>