Add method `getArgumentValue()` to return the value of the argument specified over the command line
The value over the command-line for the given argument is returned by the
function otherwise, a default value explicitly mentioned by the user is
returned if the value is not present over the command-line.

This commit also adds the following tests to test the functionality of function
"getArgumentValue()":
* testGetArgumentValueWhenNoExplicitValueIsGiven()
* testGetArgumentValueWhenExplicitValueIsGiven()

Also, this commit fixes the following issues:
* FIX: #2
* FIX: #1

Also, make sure the shift the "sysargs" array by one position in the "Argparse"
constructor to avoid looping over the program name for flag parsing.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>